The raytracer now uses WebWorkers for better performance
The raytracer now splits the canvas into multiple tiles and creates a WebWorker for each, so they can be rendered in parallel.
With this, I set the raytracer to only render a single image completly, instead of rendering "in real time". This is way better for testing and bring better results. I might bring back real time rendering at some point, but before that I want to focus on lighting and reflections.